Separate UI from main.rb
